Add support to choose the build type
Add a variable which is a list of strings which can be used to choose
which build type to use. This is defaulted to have Debug, Release and
RelWithDebInfo. This can be overridden with a list of repeated strings
from outside. Based on the build type chosen customize the build
directory mimicing the behavior of clion to generate the build data