Add support for saved settings
Add support to save the settings of the ect for a project locally in a
file with path .ect.json in the parent directory of the project. This
will be used to save things like custom source directories, last target
that was built, generator to be used for the project etc. The general
idea is to create a small config of things that can be different from
project to project and doesnt have to be setup every time a project is loaded