Add some extra support for lsp
Add support for auto configuring the parameters required for lsp. The
assumption here is that the lsp used is clangd. So this is not
generalised. If someone is using ccls or rtags, it requires some
additional changes to generalize this

Along with this by default we save the configuration when an action like
changing build type or build target is done. This is done to make sure
that on restart the correct settings are reloaded